+++ include/CTGUI/Widgets/ListBox.h
@@ -64,8 +64,8 @@ CTGUI_API unsigned int tguiListBox_getItemHeight(const tguiWidget* widget);
 CTGUI_API void tguiListBox_setTextSize(tguiWidget* widget, unsigned int size);
 CTGUI_API unsigned int tguiListBox_getTextSize(const tguiWidget* widget);
 
-CTGUI_API void tguiListBox_setMaximumItems(tguiWidget* widget, unsigned int maxItems);
-CTGUI_API unsigned int tguiListBox_getMaximumItems(const tguiWidget* widget);
+CTGUI_API void tguiListBox_setMaximumItems(tguiWidget* widget, size_t maxItems);
+CTGUI_API size_t tguiListBox_getMaximumItems(const tguiWidget* widget);
 
 CTGUI_API void tguiListBox_setAutoScroll(tguiWidget* widget, sfBool autoScroll);
 CTGUI_API sfBool tguiListBox_getAutoScroll(const tguiWidget* widget);
